i2c function calls
- read/write i2c function calls based on datasheet
- sequential read/write adaptation for vl53l0x sensor

+++ drivers/serial_uart.h
@@ -12,7 +12,7 @@
 
 typedef struct
 {
-    uint64_t fcy; 
+    uint32_t fcy; 
     uint32_t baud;
 } uart_config_t;
